Implemented RTOS, Interrupts and Error Handling
Implemented FreeRTOS for spotify api fetch task, now the api call does not block the main loop as it runs parallel to it. 
Also implemented interrupts for button presses, now no button press goes unregistered (no matter if the loop is blocked or not).
Added error handling to handle invalid or bad responses for spotify.